backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thoth-backend"
version = "1.0.0"
authors = [
  { name="Marco Pancotti", email="mp@tylconsulting.it" },
]
description = "Thoth AI Backend - Django application for AI-powered database management and analysis"
readme = "README.md"
keywords = ["django", "ai", "database", "backend", "thoth"]
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Framework :: Django",
    "Development Status :: 4 - Beta",
]
dependencies = [
    # Core Django Framework
    "Django==5.2",
    "djangorestframework==3.16.0",
    "django-allauth==65.7.0",
    "django-crispy-forms==2.3",
    "crispy-bootstrap5==2025.4",
    "django-cors-headers==4.3.1",
    # Development Tools
    "django-browser-reload==1.18.0",
    # Installer Dependencies
    "fastapi==0.115.6",
    "uvicorn==0.34.0",
    # Essential Utilities
    "python-dotenv==1.1.1",
    "PyYAML==6.0.2",
    "toml==0.10.2",
    "requests==2.32.3",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "chardet==5.2.0",
    # PDF Generation
    "weasyprint==62.3",
    "xhtml2pdf==0.2.16",
    # Data Processing
    "pandas==2.2.3",
    "tabulate==0.9.0",
    # Database Support
    "SQLAlchemy==2.0.40",
    "psycopg2-binary==2.9.10",
    # Thoth-specific Libraries (Qdrant as default vector DB)
    # Default: PostgreSQL and SQLite (no system dependencies required)
    "thoth-dbmanager[postgresql,sqlite]==0.5.8",
    # Additional SQL Server drivers for better compatibility
    "pymssql>=2.2.0",
    "thoth-qdrant==0.1.8",  # Default vector database  
    # AI/ML Framework (LiteLLM - lightweight LLM library)
    "litellm>=1.75.5",
    # External Embedding Service Clients
    "openai==1.99.9",
    "cohere>=4.0.0",
    "mistralai>=0.1.0",
    "requests>=2.25.0",
    "tenacity>=8.0.0",
    # Production Server
    "gunicorn==23.0.0",
    # Essential System Dependencies
    "certifi==2025.1.31",
    "psutil>=6.0.0",
]

[project.optional-dependencies]
# Optional database support - installa SOLO i driver aggiuntivi necessari
# Nota: PostgreSQL e SQLite sono già inclusi nelle dipendenze base
mariadb = [
    "mariadb>=1.1.0",  # Solo il driver MariaDB
]
sqlserver = [
    "pyodbc>=4.0.0",   # Solo il driver SQL Server
]
# Installa entrambi i driver opzionali
all-databases = [
    "mariadb>=1.1.0",
    "pyodbc>=4.0.0",
]

dev = [
    "pytest>=8.4.1",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "pytest-html>=3.1.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
    "safety>=2.0.0",
]

test = [
    "pytest>=8.4.1",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "pytest-html>=3.1.0",
]

[project.urls]
"Homepage" = "https://github.com/mptyl/thoth_be"
"Bug Tracker" = "https://github.com/mptyl/thoth_be/issues"
"Documentation" = "https://github.com/mptyl/thoth_be#readme"
"Source Code" = "https://github.com/mptyl/thoth_be"

[tool.setuptools.packages.find]
where = ["."]
include = ["thoth_*", "Thoth*"]
exclude = ["tests*", "test_*", "*.tests", "*.tests.*", "data*", "logs*", "exports*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.yml", "*.yaml", "*.html", "*.css", "*.js"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "Thoth.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = ["thoth_core", "thoth_ai_backend", "Thoth"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
